+++ packages/amqp/lib/connect.js
@@ -2,8 +2,6 @@ const amqp = require('amqplib');
 const safePromise = require('@idearium/safe-promise');
 const log = require('@idearium/log')();
 
-const amqpUrl = process.env.MQ_URL;
-
 const redactUrl = (url) => {
     const [protocol, remainder] = url.split('://');
     let host = remainder;
@@ -15,61 +13,45 @@ const redactUrl = (url) => {
     return `${protocol}://${host}`;
 };
 
-module.exports = () => {
-    let _connection;
-    let _state = 'disconnected';
+module.exports = async (mqUrl, opts = {}) => {
+    let state = 'disconnected';
 
-    const createConnection = async (mqUrl = amqpUrl, opts = {}) => {
-        if (!mqUrl) {
-            throw new Error('mqUrl parameter is required');
-        }
+    if (!mqUrl) {
+        throw new Error('mqUrl parameter is required');
+    }
 
-        const url = redactUrl(mqUrl);
+    const url = redactUrl(mqUrl);
 
-        log.info({ url }, 'Connecting to AMQP server.');
+    log.info({ url }, 'Connecting to AMQP server.');
 
-        _state = 'connecting';
+    state = 'connecting';
 
-        const [err, connection] = await safePromise(amqp.connect(mqUrl, opts));
+    const [err, connection] = await safePromise(amqp.connect(mqUrl, opts));
 
-        if (err) {
-            log.error({ err, url }, 'Could not connect to AMQP server.');
+    if (err) {
+        log.error({ err, url }, 'Could not connect to AMQP server.');
 
-            throw err;
-        }
+        throw err;
+    }
 
-        _state = 'connected';
+    state = 'connected';
 
-        log.info({ url }, 'Connected to AMQP server.');
+    log.info({ url }, 'Connected to AMQP server.');
 
-        // This needs to be async so that the throw causes Node.js to exit.
-        connection.on('close', async (connectionErr) => {
-            _state = 'disconnected';
+    // This needs to be async so that the throw causes Node.js to exit.
+    connection.on('close', async (connectionErr) => {
+        state = 'disconnected';
 
-            log.error(
-                { err: connectionErr, url },
-                'Connection to the AMQP server closed.'
-            );
+        log.error(
+            { err: connectionErr, url },
+            'Connection to the AMQP server closed.'
+        );
 
-            // This didn't wor for some reason?
-            throw new Error('The connection to the AMQP server closed.');
-        });
+        // This didn't wor for some reason?
+        throw new Error('The connection to the AMQP server closed.');
+    });
 
-        return connection;
-    };
+    connection.isConnected = () => state === 'connected';
 
-    const connect = (url, opts) => {
-        if (!_connection) {
-            _connection = createConnection(url, opts);
-        }
-
-        // eslint-disable-next-line no-use-before-define
-        return _connection;
-    };
-
-    return {
-        connect,
-        connection: () => _connection,
-        isConnected: () => _state === 'connected',
-    };
+    return connection;
 };