+++ packages/amqp/tests/lib/connect.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const connectTest = () => {
+    return { connect: () => {} };
+};
+
+module.exports = connectTest;